change compiling on nt

+++ nt.ps1
@@ -1,6 +1,15 @@
-param([Array]$envs=('39', '310', '311'), [String]$root='C:\Users\Chen\Projects\Python\informatics')
-$dist_dir = $root+'/dist'
-$doc_dir = $root+'/docs'
+# author: Chen Zhang
+# usage:
+#     * compiling for python 3.9 | 3.10 | 3.11 with documents embedded
+#       > nt.ps1
+#     * compiling for python 3.11 without documents embedded
+#       > nt.ps1 '311' -add_doc $FALSE
+
+
+param([Array]$envs=('39', '310', '311'), [String]$root='C:\Users\Chen\Projects\Python\informatics',
+      [Bool]$add_doc=$TRUE)
+$dist_dir = $root+'\dist'
+$doc_dir = $root+'\docs'
 
 function activate_conda_hook() {
     C:\ProgramData\anaconda3\shell\condabin\conda-hook.ps1
@@ -15,14 +24,18 @@ function single_compiler([String]$env) {
     $installed = pip list | Where-Object {$_ -Match 'informatics'}
     if ($installed) {pip uninstall informatics --yes}
     python -m build -w -Cbuild_ext
-    Set-Location $dist_dir
-    $whl1 = Get-ChildItem -Name | Where-Object {$_ -Match 'win_amd64.whl$'} | Where-Object {$_ -Match 'cp'+$env}
-    if ($whl1) {pip install $whl1}
-    Set-Location $doc_dir
-    (make clean)-and(make html)
-    Set-Location $root
-    python -m build -w -Cbuild_ext
-    pip uninstall informatics --yes
+
+    if ($add_doc) {
+        Set-Location $dist_dir
+        $whl1 = Get-ChildItem -Name | Where-Object {$_ -Match 'win_amd64.whl$'} | Where-Object {$_ -Match 'cp'+$env}
+        if ($whl1) {pip install $whl1}
+        Set-Location $doc_dir
+        (make clean)-and(make html)
+        Set-Location $root
+        python -m build -w -Cbuild_ext
+        pip uninstall informatics --yes
+    }
+
     conda deactivate
 }
